requirements.txt
Adafruit-Blinka==8.52.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-neopixel==6.3.15
adafruit-circuitpython-pixelbuf==2.0.7
adafruit-circuitpython-requests==4.1.9
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
annotated-types==0.7.0
anyio==4.8.0
binho-host-adapter==0.1.6
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
openai==1.60.2
pydantic==2.10.6
pydantic_core==2.27.2
pyftdi==0.56.0
pyserial==3.5
pyusb==1.3.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
sniffio==1.3.1
sysv-ipc==1.1.0
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
